classify_data/requirements.txt
flask
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-translate==3.19.0
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.14.0
gunicorn==23.0.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
nltk==3.9.1
numpy==2.0.2
oauthlib==3.2.2
pandas==2.2.3
PyYAML==6.0.2
regex==2024.9.11
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
tqdm==4.66.6
urllib3==2.2.3
yarl==1.18.3
pytz==2024.2
regex==2024.9.11
sentence-transformers==3.4.1
tokenizers==0.21.0
transformers==4.48.3